api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.4
SQLAlchemy==2.0.30
aiomysql==0.2.0
python-dotenv==1.0.1